Test long description;-;Getting fired sucks. It doesn't matter how or when. It just sucks. And right now there are an awful lot of people in the tech industry feeling just how much. But what's bad for the individual isn't always bad for the group. Believe it or not, there's also collective upside to the massive tech layoffs happening at the moment.
Like undoing the absurd hiring spree committed by the largest tech companies during the pandemic. This spree swallowed up a preposterous amount of talent, then put much of it to work on bullshit projects with little to no value for the company or the economy as a whole.

Such hoarding of tech talent represented a tragic misallocation of productive capacity. The likes of Amazon, Meta, Google, and many others added tens of thousands of positions to already bloated bureaucracies, which often just made them slower and less capable.

Not because the people they hired weren't smart, caring, or capable people, mind you! But because there's only so much brain power that can productively be thrown at optimizing ad clicks or poured into VR black holes or operating ever-more authoritarian censorship regimes.

Now this might seem like a "not-a-me problem" from the perspective of highly-paid tech workers grinding out RSUs in the belly of some overgrown, monopolistic beast. Who cares whether investors see the best return on their capital or whether society at large could have used the capacity better. There's a smoothie bar here and free sushi for lunch!

But I'd argue that as much as getting fired sucks, and that's a lot, it's at least an immediate, explicit pain. Unlike the slow, creeping pain of having wasted your talent on a bullshit job that the world would have been oblivious to be without. The kind of existential wound that David Graeber tried to bandage with Bullshit Jobs.

Now arguing what is or isn't best for the individual is tricky business. I'm sure there are plenty of laid off tech workers who'd say they'd happily hold that bullshit job if it meant paying the mortgage. I get that. So let's return to the group perspective.

When these massive tech companies hoard all the talent, it goes without saying that there's less of it to go around. Demand might produce supply, but there's a long lag factor before it does when education is involved. So all sorts of companies across the economy have found themselves priced out of access to capable tech workers. Meaning that all sorts of less glamorous, but sorely needed businesses were left stranded, and in return made all of us worse off.

I hope everyone who've been laid off in this tsunami of firings find another job quickly. But I also hope that those jobs are with a broader set of companies. From startups to stayups. Maybe it won't quite all be at the same gold-rush wages, but it'll hopefully be in a more productive, fulfilling capacity.

May the three hundred thousand plus brains that have been released so far bloom a thousand new fields!